+++ CHANGELOG.md
@@ -5,6 +5,15 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 ## [Unreleased]
 ### Added
+
+### Changed
+
+### Fixed
+
+### Removed
+
+## [v3.5.0] - 2023-06-09
+### Added
 - added bind command to meter feature to myems-api and myems-admin
 - added bind command to space feature to myems-api and myems-admin
 - added forgot pssword function to myems-api and myems-web
@@ -13,8 +22,13 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 - added bind command to equipment feature to myems-api and myems-admin
 - added bind command to combined equipment feature to myems-api and myems-admin
 - added new user approval function to myems-admin
-
+- added bind command to tenant feature to myems-api and myems-admin
+- added bind command to store feature to myems-api and myems-admin
+- added bind command to shopfloor feature to myems-api and myems-admin
+- added new tables for microgrids to system database
+- 
 ### Changed
+- updated the default tariff type and default tariff valid through date time in tariff editor
 
 ### Fixed
 - fixed save button issue of tariff dialog in myems-admin
@@ -1300,7 +1314,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 ### Removed
 -   None.
 
-[Unreleased]: https://github.com/MyEMS/myems/compare/v3.4.0...HEAD
+[Unreleased]: https://github.com/MyEMS/myems/compare/v3.5.0...HEAD
+[3.5.0]: https://github.com/MyEMS/myems/compare/v3.5.0...v3.4.0
 [3.4.0]: https://github.com/MyEMS/myems/compare/v3.4.0...v3.3.0
 [3.3.0]: https://github.com/MyEMS/myems/compare/v3.3.0...v3.2.0
 [3.2.0]: https://github.com/MyEMS/myems/compare/v3.1.0...v3.2.0